backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
audioread==3.0.1
bcrypt==4.1.2
cachetools==5.3.3
captain==4.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
datatypes==0.12.0
decorator==5.1.1
docopt==0.6.2
ecdsa==0.18.0
fastapi==0.110.1
filelock==3.14.0
fsspec==2024.5.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-speech==2.26.0
google-cloud-storage==2.16.0
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
idna==3.6
inflect==7.2.0
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonpickle==0.9.6
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
MarkupSafe==2.1.5
mido==1.3.2
mir_eval @ git+https://github.com/craffel/mir_eval.git@485a42551f449517567313f9871f58e9c99fbd8a
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
munch==2.5.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
packaging==23.2
passlib==1.7.4
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
proto-plus==1.23.0
protobuf==4.25.3
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
PyMySQL==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.2
rsa==4.9
sacred==0.7.4
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.5.0
torch==2.3.0
tqdm==4.66.4
typeguard==4.2.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
wrapt==1.16.0
